+++ src/test/java/com/woowacamp/soolsool/global/auth/util/TokenProviderIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.global.auth;
+package com.woowacamp.soolsool.global.auth.util;
 
 import static com.woowacamp.soolsool.core.member.code.MemberErrorCode.MEMBER_NO_INFORMATION;
 import static com.woowacamp.soolsool.global.auth.code.AuthErrorCode.TOKEN_ERROR;
@@ -6,7 +6,6 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import com.woowacamp.soolsool.core.member.domain.Member;
 import com.woowacamp.soolsool.core.member.repository.MemberRepository;
-import com.woowacamp.soolsool.global.auth.util.TokenProvider;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -19,7 +18,8 @@ import org.springframework.context.annotation.Import;
 
 @DataJpaTest
 @Import(TokenProvider.class)
-class TokenProviderTest {
+@DisplayName("TokenProvider 통합 테스트")
+class TokenProviderIntegrationTest {
 
     @Autowired
     private TokenProvider tokenProvider;
